pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ffmpeg-benchmark"
version = "0.2.1"
authors = [
    { name = "Anthony Monthe", email = "anthony@cloud-mercato.com" },
]
description = "Handy tool for evaluate ffmpeg performance"
readme = "README.rst"
license = { text = "MIT" }
requires-python = ">=3"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: File Transfer Protocol (FTP)",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Networking",
    "Topic :: System :: Benchmark",
]
keywords = ["benchmark", "machine learning", "deep learning"]
dependencies = [
    "ffmpeg-python>=0.2.0",
    "handystats @ https://github.com/cloudmercato/handystats/archive/refs/heads/main.zip",
    "probes",
]

[project.urls]
Homepage = "https://github.com/cloudmercato/ffmpeg-benchmark"

[project.scripts]
ffmpeg-benchmark = "ffmpeg_benchmark.main:main"

[tool.hatch.build]
# This is a basic configuration. You might need to adjust it based on your project structure.
# By default, hatchling includes everything in the src directory if it exists,
# or the root directory if src does not exist.
# You can customize this using include/exclude patterns.
# For example:
# include = ["/ffmpeg_benchmark", "/other_package"] # if your packages are in the root
# include = ["/src/ffmpeg_benchmark"] # if your package is in src
# exclude = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.coverage.run]
omit = ["ffmpeg_benchmark/tests/*"]
branch = true
source = ["ffmpeg_benchmark"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "noqa",
    "NotImplemented",
]

[tool.uv.sources]
probes = { git = "https://github.com/cloudmercato/Probes" }

[dependency-groups]
dev = [
    "pyright>=0.0.13.post0",
]
